Retravaillé adjonction et retrait de produits à une commande

+++ assets/BDR - EasyService StoredProcedures.sql
@@ -66,6 +66,7 @@ CREATE PROCEDURE CommandeAjouterProduit (
 BEGIN
 
 	DECLARE OldNbrProduits INT UNSIGNED;
+    DECLARE AdditionNum INT UNSIGNED;
     
     DECLARE EXIT HANDLER FOR SQLEXCEPTION -- garantit l'atomicité de la transaction
     BEGIN
@@ -73,25 +74,32 @@ BEGIN
         RESIGNAL;
     END;
     
-	START TRANSACTION;
-		SELECT nbrDeProduit
-			INTO OldNbrProduits
-			FROM Commande_Produit
-			WHERE idCommande = inIdCommande
-				AND idProduit = inIdProduit;
-				
-		IF OldNbrProduits IS NULL
-		THEN
-			INSERT INTO Commande_Produit (idCommande, idProduit, nbrDeProduit, sortiDeCuisine)
-			VALUES (inIdCommande, inIdProduit, inNbrProduits, 0);
-		ELSE
-			UPDATE Commande_Produit
-			SET nbrDeProduit = oldNbrProduits + inNbrProduits
-			WHERE 
-				idCommande = inIdCommande
-				AND idProduit = inIdProduit;
-		END IF;
-    COMMIT;
+    SELECT idAddition
+    INTO AdditionNum
+    FROM Commande
+    WHERE id = inIdCommande;
+    
+    IF AdditionNum IS NULL THEN
+		START TRANSACTION;
+			SELECT nbrDeProduit
+				INTO OldNbrProduits
+				FROM Commande_Produit
+				WHERE idCommande = inIdCommande
+					AND idProduit = inIdProduit;
+					
+			IF OldNbrProduits IS NULL
+			THEN
+				INSERT INTO Commande_Produit (idCommande, idProduit, nbrDeProduit, sortiDeCuisine)
+				VALUES (inIdCommande, inIdProduit, inNbrProduits, 0);
+			ELSE
+				UPDATE Commande_Produit
+				SET nbrDeProduit = oldNbrProduits + inNbrProduits
+				WHERE 
+					idCommande = inIdCommande
+					AND idProduit = inIdProduit;
+			END IF;
+		COMMIT;
+    END IF;
 END //
 
 DROP PROCEDURE IF EXISTS CommandeRetirerProduit;
@@ -104,6 +112,7 @@ BEGIN
 
 	DECLARE MustRemoveAll TINYINT;
 	DECLARE NumProduitRestant INT UNSIGNED;
+    DECLARE AdditionNum INT UNSIGNED;
     
     DECLARE EXIT HANDLER FOR SQLEXCEPTION -- garantit l'atomicité de la transaction
     BEGIN
@@ -111,36 +120,43 @@ BEGIN
         RESIGNAL;
     END;
     
-    START TRANSACTION;
-		SELECT 
-			CASE
-				WHEN inNbrDeProduit < nbrDeProduit THEN FALSE
-				WHEN inNbrDeProduit = nbrDeProduit THEN TRUE
-			END AS mustRemove
-		INTO MustRemoveAll
-		FROM Commande_Produit
-		WHERE idCommande = inIdCommande
-			AND idProduit = inIdProduit;
-			
-		IF MustRemoveAll = TRUE
-			THEN
-				DELETE FROM Commande_Produit
-				WHERE idCommande = inIdCommande
-					AND idProduit = inIdProduit;
-			ELSE
-				SELECT (nbrDeProduit - inNbrDeProduit)
-				INTO numProduitRestant
-				FROM Commande_Produit
-				WHERE idCommande = inIdCommande
-					AND idProduit = inIdProduit;
+    SELECT idAddition
+    INTO AdditionNum
+    FROM Commande
+    WHERE id = inIdCommande;
+    
+    IF AdditionNum IS NULL THEN
+		START TRANSACTION;
+			SELECT 
+				CASE
+					WHEN inNbrDeProduit < nbrDeProduit THEN FALSE
+					WHEN inNbrDeProduit = nbrDeProduit THEN TRUE
+				END AS mustRemove
+			INTO MustRemoveAll
+			FROM Commande_Produit
+			WHERE idCommande = inIdCommande
+				AND idProduit = inIdProduit;
 				
-				
-				UPDATE Commande_Produit
-				SET nbrDeProduit = NumProduitRestant
-				WHERE idCommande = inIdCommande
-					AND idProduit = inIdProduit;
-		END IF;
-    COMMIT;
+			IF MustRemoveAll = TRUE
+				THEN
+					DELETE FROM Commande_Produit
+					WHERE idCommande = inIdCommande
+						AND idProduit = inIdProduit;
+				ELSE
+					SELECT (nbrDeProduit - inNbrDeProduit)
+					INTO numProduitRestant
+					FROM Commande_Produit
+					WHERE idCommande = inIdCommande
+						AND idProduit = inIdProduit;
+					
+					
+					UPDATE Commande_Produit
+					SET nbrDeProduit = NumProduitRestant
+					WHERE idCommande = inIdCommande
+						AND idProduit = inIdProduit;
+			END IF;
+		COMMIT;
+    END IF;
 END //
 
 DROP PROCEDURE IF EXISTS AdditionCreer;
@@ -151,41 +167,53 @@ BEGIN
 
 	DECLARE InsertedId INT UNSIGNED;
     DECLARE Prix FLOAT;
+    DECLARE AdditionNum INT UNSIGNED;
 	
     DECLARE EXIT HANDLER FOR SQLEXCEPTION -- garantit l'atomicité de la transaction
     BEGIN
         ROLLBACK; 
         RESIGNAL;
     END;
+    SELECT idAddition
+    INTO additionNum
+    FROM Commande
+    WHERE id = inIdCommande;
     
-	START TRANSACTION;
-		INSERT INTO Log (dateLog)
-		VALUES (NOW());
-		
-		SELECT LAST_INSERT_ID()
-		INTO InsertedId;
-		
-		SELECT SUM(prixVente) 
-		INTO Prix
-		FROM Produit
-		JOIN Commande_Produit ON Commande_Produit.idProduit = Produit.id;
-		
-		INSERT INTO Addition (idLog, coutPrixTotal, estPaye, idCommande)
-		VALUES (
-			InsertedId,
-			Prix, 
-			FALSE,
-			inIdCommande);
-	COMMIT;
+    IF AdditionNum IS NULL THEN
+		START TRANSACTION;
+			INSERT INTO Log (dateLog)
+			VALUES (NOW());
+			
+			SELECT LAST_INSERT_ID()
+			INTO InsertedId;
+			
+			SELECT SUM(prixVente) 
+			INTO Prix
+			FROM Produit
+			JOIN Commande_Produit ON Commande_Produit.idProduit = Produit.id;
+			
+			INSERT INTO Addition (idLog, coutPrixTotal, estPaye)
+			VALUES (
+				InsertedId,
+				Prix, 
+				FALSE);
+			
+			UPDATE Commande
+			SET idAddition = InsertedId
+			WHERE id = inIdCommande;
+			
+		COMMIT;
+	END IF;
 END //
 
 DROP PROCEDURE IF EXISTS AdditionPayer;
 DELIMITER //
 CREATE PROCEDURE AdditionPayer (
-	IN inIdAddition INT UNSIGNED)
+	IN inIdCommande INT UNSIGNED)
 BEGIN
 	
-    DECLARE TableSalleId INT UNSIGNED;
+    DECLARE TableSalleId INT UNSIGNED DEFAULT 0;
+    DECLARE IdUpdatedAddition INT UNSIGNED DEFAULT 0;
     
     DECLARE EXIT HANDLER FOR SQLEXCEPTION -- garantit l'atomicité de la transaction
     BEGIN
@@ -194,16 +222,22 @@ BEGIN
     END;
     
 	START TRANSACTION;
+    
+		SELECT idAddition
+        INTO IdUpdatedAddition
+        FROM Commande
+        WHERE id = inIdCommande;
+		
 		SELECT idTableSalle
 		INTO TableSalleId
 		FROM Commande
-		WHERE idAddition = inIdAddition;
+		WHERE idAddition = IdUpdatedAddition;
 		
-		UPDATE addition
+		UPDATE Addition
 		SET estPaye = TRUE
-		WHERE idLog = inIdAddition;
+		WHERE idLog = IdUpdatedAddition;
 		
-		UPDATE tablesalle
+		UPDATE TableSalle
 		SET ouvertFerme = FALSE
 		WHERE id = TableSalleId;
 	COMMIT;
@@ -295,7 +329,7 @@ BEGIN
     
 	START TRANSACTION;
 	
-    SELECT COUNT(idArticleStock) - 1
+    SELECT COUNT(idArticleStock)
     INTO MaxForLoop
     FROM ArticleStock_Produit
     WHERE idProduit = inIdProduit;
